+++ CHANGELOG.md
@@ -19,3 +19,11 @@ it's contained:
 1. Automatically complete daily tasks(including PK, 30 boxes brushing).
 2. Automatic card drawing.
 3. Automatic wood brushing (5-2 levels).
+
+2025-05-25
+\- [improve]\(001\): improved some daily task codes
+
+it's contained: 
+1. Daily Shopping Gift Pack
+2. Daily Hero Gift
+3. Daily check-in